omnirose/requirements.txt
django[bcrypt]>=1.8,<1.9
psycopg2
numpy
scipy
cairocffi
django-template-email
django-bootstrap-form
selenium
requests
stripe
django-settings-export
python-postmark
